Updating rendering of Badge
